+++ src/Codeception/Module/ZendExpressive.php
@@ -3,7 +3,6 @@ namespace Codeception\Module;
 
 use Codeception\Lib\Framework;
 use Codeception\TestInterface;
-use Codeception\Configuration;
 use Codeception\Lib\Connector\ZendExpressive as ZendExpressiveConnector;
 use Codeception\Lib\Interfaces\DoctrineProvider;
 
@@ -19,9 +18,11 @@ use Codeception\Lib\Interfaces\DoctrineProvider;
  *
  * ## Config
  *
- * * container: relative path to file which returns Container (default: `config/container.php`)
+ * * `container` - (default: `config/container.php`) relative path to file which returns Container
+ * * `recreateApplicationBetweenTests` - (default: false) whether to recreate the whole application before each test
+ * * `recreateApplicationBetweenRequests` - (default: false) whether to recreate the whole application before each request
  *
- * ## API
+ * ## Public properties
  *
  * * application -  instance of `\Zend\Expressive\Application`
  * * container - instance of `\Interop\Container\ContainerInterface`
@@ -31,7 +32,9 @@ use Codeception\Lib\Interfaces\DoctrineProvider;
 class ZendExpressive extends Framework implements DoctrineProvider
 {
     protected $config = [
-        'container' => 'config/container.php',
+        'container'                          => 'config/container.php',
+        'recreateApplicationBetweenTests'    => false,
+        'recreateApplicationBetweenRequests' => false,
     ];
 
     /**
@@ -41,55 +44,37 @@ class ZendExpressive extends Framework implements DoctrineProvider
 
     /**
      * @var \Interop\Container\ContainerInterface
+     * @deprecated Doesn't work as expected if Application is recreated between requests
      */
     public $container;
 
     /**
      * @var \Zend\Expressive\Application
+     * @deprecated Doesn't work as expected if Application is recreated between requests
      */
     public $application;
 
-    protected $responseCollector;
-
     public function _initialize()
     {
-        $cwd = getcwd();
-        $projectDir = Configuration::projectDir();
-        chdir($projectDir);
-        $this->container = require $projectDir . $this->config['container'];
-        $app = $this->container->get(\Zend\Expressive\Application::class);
+        $this->client = new ZendExpressiveConnector();
+        $this->client->setConfig($this->config);
 
-        $middlewareFactory = null;
-        if ($this->container->has(\Zend\Expressive\MiddlewareFactory::class)) {
-            $middlewareFactory = $this->container->get(\Zend\Expressive\MiddlewareFactory::class);
+        if ($this->config['recreateApplicationBetweenTests'] == false && $this->config['recreateApplicationBetweenRequests'] == false) {
+            $this->application = $this->client->initApplication();
+            $this->container   = $this->client->getContainer();
         }
-
-        $pipelineFile = $projectDir . 'config/pipeline.php';
-        if (file_exists($pipelineFile)) {
-            $pipelineFunction = require $pipelineFile;
-            if (is_callable($pipelineFunction) && $middlewareFactory) {
-                $pipelineFunction($app, $middlewareFactory, $this->container);
-            }
-        }
-        $routesFile = $projectDir . 'config/routes.php';
-        if (file_exists($routesFile)) {
-            $routesFunction = require $routesFile;
-            if (is_callable($routesFunction) && $middlewareFactory) {
-                $routesFunction($app, $middlewareFactory, $this->container);
-            }
-        }
-        chdir($cwd);
-
-        $this->application = $app;
-        $this->initResponseCollector();
     }
 
     public function _before(TestInterface $test)
     {
         $this->client = new ZendExpressiveConnector();
-        $this->client->setApplication($this->application);
-        if ($this->responseCollector) {
-            $this->client->setResponseCollector($this->responseCollector);
+        $this->client->setConfig($this->config);
+
+        if ($this->config['recreateApplicationBetweenTests'] != false && $this->config['recreateApplicationBetweenRequests'] == false) {
+            $this->application = $this->client->initApplication();
+            $this->container   = $this->client->getContainer();
+        } elseif (isset($this->application)) {
+            $this->client->setApplication($this->application);
         }
     }
 
@@ -101,25 +86,6 @@ class ZendExpressive extends Framework implements DoctrineProvider
         }
 
         parent::_after($test);
-    }
-
-    private function initResponseCollector()
-    {
-        if (!method_exists($this->application, 'getEmitter')) {
-            //Does not exist in Zend Expressive v3
-            return;
-        }
-
-        /**
-         * @var Zend\Expressive\Emitter\EmitterStack
-         */
-        $emitterStack = $this->application->getEmitter();
-        while (!$emitterStack->isEmpty()) {
-            $emitterStack->pop();
-        }
-
-        $this->responseCollector = new ZendExpressiveConnector\ResponseCollector;
-        $emitterStack->unshift($this->responseCollector);
     }
 
     public function _getEntityManager()
